+++ src/go.mod
@@ -1,0 +1,3 @@
+module pets
+
+go 1.23.3
