--- src/pathExpression/go.mod
@@ -1,3 +1,0 @@
-module pe
-
-go 1.23.3
